Render pages using the parser w/o .yaml extension
Assuming a file app/views/pages/first-post.html.erb.yaml with the
contents:

    ---
    title: My first post!
    type: post
    ---

    <p>I can't wait to see where this <%= @post.type %> goes.</p>

We should use the renderer that would be used for a hypothetical file
app/views/pages/first-post.html.erb, and that shouldn't have any of the
frontmatter included when it is rendered.

We can make this happen by parsing the original file as YAML - the
default parser will see the closing `---` and assume that a new document
has started / the current document is finished. We can use this to find
the last line of the frontmatter and split the original file into
frontmatter, discarded by the template handler, and the source, passed
to the handler based on the filename without .yaml / .yml.

Separately, the class `Post < ActionPage::Base` will provide the
frontmatter as data, so the `@post.type` will be available assuming a
controller that provides `@post = Post.find("first-post")`.